requirements.txt
pytest
pytest-cov
pylint
fire
black
streamlit
ipython
pypdf
langchain
chromadb # Vector storage
groq
langchain-core
langchain-groq
gradio
langchain_community
dvc
dvc-gdrive
ipykernel
streamlit-authenticator
python-dotenv
